+++ api/src/assets/Lists/specialitiesList.ts
@@ -1,0 +1,19 @@
+const specialitiesList: any = [
+  '{"name":"Pedriatry"}',
+  '{"name":"Obstetrician"}',
+  '{"name":"Cardiologist"}',
+  '{"name":"Oncologist"}',
+  '{"name":"Gastroenterologist"}',
+  '{"name":"Pulmonologist"}',
+  '{"name":"Nephrologist"}',
+  '{"name":"Endocrinologist"}',
+  '{"name":"Ophthalmologist"}',
+  '{"name":"Otolaryngologist"}',
+  '{"name":"Dermatologist"}',
+  '{"name":"Psychiatrist"}',
+  '{"name":"Neurologist"}',
+  '{"name":"Radiologist"}',
+  '{"name":"Allergy and Immunology"}',
+];
+
+export default specialitiesList;
