+++ api/src/assets/Lists/studiesList.ts
