+++ api/src/assets/populateSpecialities.ts
@@ -1,6 +1,6 @@
 import { Specialitie } from "../models/Specialitie";
 
-export default async function populateDB(params: []) {
+export default async function populateSpecialities(params: []) {
   const dataFromDB = await Specialitie.findAll();
   if (dataFromDB.length === 0) {
     let arrParsed: Array<any> = [];
