--- api/src/assets/specialitiesJSON.ts
@@ -1,19 +1,0 @@
-const jsonSpecialities: any = [
-    '{"name":"Pedriatry"}',
-    '{"name":"Obstetrician"}',
-    '{"name":"Cardiologist"}',
-    '{"name":"Oncologist"}',
-    '{"name":"Gastroenterologist"}',
-    '{"name":"Pulmonologist"}',
-    '{"name":"Nephrologist"}',
-    '{"name":"Endocrinologist"}',
-    '{"name":"Ophthalmologist"}',
-    '{"name":"Otolaryngologist"}',
-    '{"name":"Dermatologist"}',
-    '{"name":"Psychiatrist"}',
-    '{"name":"Neurologist"}',
-    '{"name":"Radiologist"}',
-    '{"name":"Allergy and Immunology"}'
-]
-
-export default jsonSpecialities
